Group management view/routes and matchmaking view/routes complete

+++ server/routes/group.js
@@ -43,6 +43,7 @@ router.get('/', function(req, res, next) {
   Given a student (cid), course (gencode) and a group-id (id), enlists a student in a group.
   Returns 200 if successful, else 500.
 */ //TODO: Check that user isn't already member of group in given course
+   //TODO: Delete group invites for user joining group?
 router.post('/join', function(req, res, next) {
   var loggedIn = isLoggedIn(req, res);
   if(loggedIn){
@@ -100,22 +101,41 @@ router.get('/members', function(req, res, next) {
   // have to look into how to make join querys. Worst case scenario we just write it out.
 });
 
+router.get('/invite', function(req, res, next){
+  console.log("THE DANK");
+  var loggedIn = isLoggedIn(req, res);
+  console.log("THE STANK");
+  if(loggedIn){
+    console.log("Got into if-case");
+    var decodedToken = jwt.decode(loggedIn.token, constants.secret);
+    GroupInvite.findAll({where:{
+      receiver: decodedToken.user
+    }}).then(function(invites){
+      console.log("Call to groupinvite returned; " + invites);
+      res.json({result: 'success', invites: invites, token:loggedIn.token});
+    });
+  }
+});
 /*
   Given a student (cid), course (gencode) and a group-id (id), de-lists a student from a group.
   Returns 200 upon successful de-listment, otherwise 500.
 */
 router.post('/leave', function(req, res, next) {
-  Group.destroy({
-    where: {
-      id: parseInt(req.body.id),
-      student: req.body.student,
-      course: req.body.course
-    }
-  }).then(function() {
-    res.status(200).send("deleted properly");
-  }).catch(function(err) {
-    res.status(500).send("something went wrong while leaving the group");
-  });
+  var loggedIn = isLoggedIn(req, res);
+  if(loggedIn){
+    Group.destroy({
+      where: {
+        id: parseInt(req.body.id),
+        student: req.body.student,
+        course: req.body.course
+      }
+    }).then(function() {
+      res.json({result: "success", token: loggedIn.token});
+    }).catch(function(err) {
+      res.status(500).send("something went wrong while leaving the group");
+    });
+  }
+
 });
 
 router.post('/invite', function(req, res, next){
@@ -136,6 +156,54 @@ router.post('/invite', function(req, res, next){
         });
       }else{
         res.json({result: "success", message: "Invite already sent to this user", token: loggedIn.token});
+      }
+    });
+  }
+});
+
+router.put('/invite', function(req, res, next){
+  console.log("PUT DA INVITE ON DE LEFT HAND SIDE");
+  var loggedIn = isLoggedIn(req, res);
+  if(loggedIn){
+    var decodedToken = jwt.decode(loggedIn.token, constants.secret);
+    GroupInvite.findOne().then(function(invite){
+      if(invite){
+        GroupInvite.destroy({where:
+          {$or:
+            [{$or:
+              [
+               {$and: [{course: req.body.course}, {receiver: decodedToken.user}]},
+               {$and: [{course: req.body.course}, {sender: decodedToken.user}]}
+              ]
+            },
+            {$or:
+             [
+               {$and: [{course: req.body.course}, {receiver: req.body.sender}]},
+               {$and: [{course: req.body.course}, {sender: req.body.sender}]}
+             ]}
+        ] }});
+        Group.max('id',
+          {where: {course: req.body.course}})
+        .then(function(maxId){
+          var groupNo = maxId + 1;
+          Group.build({
+            id: groupNo,
+            student: decodedToken.user,
+            course: req.body.course
+          }).save().then(function() {
+            Group.build({
+              id: groupNo,
+              student: req.body.sender,
+              course: req.body.course
+            }).save().then(function(){
+              res.json({result: "success", token: loggedIn.token});
+            });
+          }).catch(function(err) {
+            res.status(500).send((err + "\n"));
+          });
+        });
+      }else{
+        res.status(404).send(("No such invite found"));
       }
     });
   }
